add the string to create user groups
