+++ functions/custom_dataframes.py
@@ -3,10 +3,10 @@ from pyspark.sql import functions as F
 from time import time
 
 
-from functions.dataframe_helpers import create_sha_values, create_catalog_schema, create_catalog_table, create_catalog_table_owner, apply_scd_type_2, assign_signalling_id
+from functions.dataframe_helpers import create_sha_values, apply_scd_type_2, assign_signalling_id
+from functions.sql_string_builders import create_catalog_table, create_catalog_table_owner, create_catalog_schema
 from functions.data_quality_functions import calculate_signalling_issues, calculate_blocking_issues
 from functions.signalling_rules import signalling_checks_dictionary
-# from functions.tables import get_table_definition
 from functions.database import get_table_definition
 from functions.data_readers import DataReader
