+++ functions/dataframe_helpers.py
@@ -1,9 +1,7 @@
 import re
 import pyspark.sql.functions as F
-from pyspark.sql import SparkSession, DataFrame
+from pyspark.sql import DataFrame
 from pyspark.sql.window import Window
-import pyspark.sql.types as T
-from pyspark.sql.functions import udf
 
 
 def create_map_column(dataframe: DataFrame, dataframe_name: str) -> DataFrame:
@@ -11,7 +9,6 @@ def create_map_column(dataframe: DataFrame, dataframe_name: str) -> DataFrame:
     Creates a new column in the dataframe with a map containing the dataframe name as the key and the 'tiltRecordID' column as the value.
 
     Args:
-        spark_session (SparkSession): The Spark session.
         dataframe (DataFrame): The input dataframe.
         dataframe_name (str): The name of the dataframe.
 
@@ -30,7 +27,6 @@ def create_sha_values(data_frame: DataFrame, col_list: list) -> DataFrame:
     Creates SHA values for the specified columns in the DataFrame.
 
     Args:
-        spark_session (SparkSession): The SparkSession object.
         data_frame (DataFrame): The input DataFrame.
         col_list (list): The list of column names to create SHA values for.
 
@@ -99,103 +95,6 @@ def clean_column_names(data_frame: DataFrame) -> DataFrame:
             new_col_name = re.sub(r"\s+", "_", new_col_name)
             data_frame = data_frame.withColumnRenamed(col, new_col_name)
     return data_frame
-
-
-def create_catalog_table(table_name: str, schema: dict) -> str:
-    """
-    Creates a SQL string to recreate a table in Delta Lake format.
-
-    This function constructs a SQL string that can be used to create a table in Delta Lake format.
-    The table is created with the provided name and schema. If the schema includes a partition column,
-    the table is partitioned by that column.
-
-    Parameters
-    ----------
-    table_name : str
-        The name of the table to be created.
-    schema : dict
-        The schema of the table to be created. The schema should be a dictionary with a 'columns' key
-        containing a list of dictionaries, each representing a column. Each column dictionary should
-        have 'name', 'type', and 'nullable' keys. The schema can optionally include a 'partition_column'
-        key with the name of the column to partition the table by.
-
-    Returns
-    -------
-    str
-        A SQL string that can be used to create the table in Delta Lake format.
-    """
-
-    if not schema["columns"]:
-        raise ValueError("The provided schema can not be empty")
-
-    create_catalog_table_string = ""
-
-    # Build a SQL string to recreate the table in the most up to date format
-    create_catalog_table_string = f"CREATE TABLE IF NOT EXISTS {table_name} ("
-
-    for i in schema["columns"]:
-        col_info = i.jsonValue()
-        print(col_info)
-        # If column is an array, for example
-        if isinstance(col_info["type"], dict):
-            col_type_parent = col_info["type"]["type"]
-            col_type_child = col_info["type"]["elementType"]
-
-            col_type = f"{col_type_parent}<{col_type_child}>"
-        else:
-            col_type = col_info["type"]
-
-        col_string = f"`{col_info['name']}` {col_type} {'NOT NULL' if not col_info['nullable'] else ''},"
-        create_catalog_table_string += col_string
-
-    create_catalog_table_string = create_catalog_table_string[:-1] + ")"
-    create_catalog_table_string += " USING DELTA "
-    if schema["partition_column"]:
-        create_catalog_table_string += (
-            f"PARTITIONED BY (`{schema['partition_column']}` STRING)"
-        )
-
-    return create_catalog_table_string
-
-
-def create_catalog_schema(environment: str, schema: dict) -> str:
-    """
-    Creates a catalog schema if it doesn't already exist.
-
-    Args:
-        environment (str): The environment in which the schema should be created.
-        schema (dict): A dictionary containing the schema details, including the container name.
-
-    Returns:
-        str: The SQL string for creating the catalog schema.
-    """
-
-    create_catalog_schema_string = (
-        f'CREATE SCHEMA IF NOT EXISTS {environment}.{schema["container"]};'
-    )
-    create_catalog_schema_owner = (
-        f'ALTER SCHEMA {environment}.{schema["container"]} SET OWNER TO tiltDevelopers;'
-    )
-
-    return create_catalog_schema_string, create_catalog_schema_owner
-
-
-def create_catalog_table_owner(table_name: str) -> str:
-    """
-    Creates a SQL string to set the owner of a table.
-
-    Args:
-        table_name (str): The name of the table.
-
-    Returns:
-        str: The SQL string to set the owner of the table.
-    """
-
-    create_catalog_table_owner_string = (
-        f"ALTER TABLE {table_name} SET OWNER TO tiltDevelopers"
-    )
-
-    return create_catalog_table_owner_string
 
 
 def apply_scd_type_2(new_table: DataFrame, existing_table: DataFrame) -> DataFrame:
